Pizza-Ban Project Business
With a solid plan in place, I proceed to execute the project. This involves coding, designing, experimenting, or whatever is necessary for bringing my ideas to life. Along the way, I might encounter challenges, which require problem-solving skills and adaptability. Regular evaluation and iteration are essential during this phase to track progress and make necessary adjustments.

The final goal is to successfully complete the project, ensuring that it meets the initial objectives and demonstrates a practical application of the knowledge gained in class. Through this process, I not only develop technical skills but also gain a deeper understanding of the subject matter, fostering independence and problem-solving abilities that will be valuable throughout my academic journey and beyond.